consolidate header map tests
tests that apply to any header map
are consolidated into AbstractHeaderMapTest

individual additional tests can go in extensions
but so far, that's not needed.
